+++ src/pages/Admin/AdminInfaq.vue
@@ -6,6 +6,7 @@ import $ from "jquery";
 import DataTable from "datatables.net-vue3";
 import Modal from "../../components/Modal.vue";
 import Swal from "sweetalert2";
+import * as XLSX from "xlsx";
 
 DataTable.use(DataTablesCore);
 
@@ -77,26 +78,41 @@ const addInfaq = async () => {
 };
 
 const deleteInfaq = async (InfaqId) => {
-  Swal.showLoading();
-
-  try {
-    const response = await fetch(
-      `${import.meta.env.VITE_BASE_URL}/api/infaq/${InfaqId}`,
-      {
-        method: "DELETE",
-        headers: {
-          Authorization: `Bearer ${token}`,
-        },
+   Swal.fire({
+    title: "Apakah Anda yakin?",
+    text: "Data yang dihapus tidak dapat dikembalikan!",
+    icon: "warning",
+    showCancelButton: true,
+   
+    confirmButtonText: "Ya, hapus!", 
+  }).then(async (result) => {
+    if (result.isConfirmed) {
+      Swal.showLoading();
+      try {
+        const response = await fetch(
+          `${import.meta.env.VITE_BASE_URL}/api/infaq/${InfaqId}`,
+          {
+            method: "DELETE",
+            headers: {
+              Authorization: "Bearer " + token,
+            },
+          }
+        );
+        if (!response.ok) {
+          throw new Error("Network response was not ok");
+        }
+        fetchInfaq();
+        Swal.close();
+        Swal.fire({
+          icon: "success",
+          title: "Berhasil",
+          text: "Data Infaq berhasil dihapus",
+        });
+      } catch (error) {
+        console.error("There was a problem deleting the infaq:", error);
       }
-    );
-    if (!response.ok) {
-      throw new Error("Network response was not ok");
     }
-    fetchInfaq();
-    Swal.close();
-  } catch (error) {
-    console.error("There was a problem deleting the Infaq:", error);
-  }
+  });
 };
 
 const initializeDataTable = () => {
@@ -159,6 +175,13 @@ const cc = [
     },
   },
 ];
+
+const exportToExcel = () => {
+  const worksheet = XLSX.utils.json_to_sheet(Infaq.value);
+  const workbook = XLSX.utils.book_new();
+  XLSX.utils.book_append_sheet(workbook, worksheet, "Infaq");
+  XLSX.writeFile(workbook, "Infaq.xlsx");
+};
 
 onMounted(() => {
   fetchInfaq();
@@ -268,19 +291,11 @@ const showModal = () => {
         <!-- Modal Create -->
 
         <div class="d-flex justify-content-end gap-3">
+          
           <button
             type="button"
             class="btn btn-primary"
-            data-toggle="modal"
-            data-target="#myModal"
-          >
-            Request
-          </button>
-          <button
-            type="button"
-            class="btn btn-primary"
-            data-toggle="modal"
-            data-target="#myModal"
+         @click="exportToExcel"
           >
             Print
           </button>
